feat: implement EpochsHistoryState as separate module for tracking pool epoch data
- Introduced EpochsHistoryState to manage and store epoch-related data for pool operators, including active stake, delegators count, and rewards.
- Added methods to handle SPOStakeDistributionMessage, SPORewardsMessage, and EpochActivityMessage, updating the epoch history accordingly.
- Integrated EpochsHistoryState into the SPOState, ensuring that epoch data is updated during relevant subscriptions.
- Created unit tests to validate the functionality of the new epoch history management.
